[Fix] Moved code from src/subdirectory, not supported by Rcpp
