+++ src/PeakIntegrator.cpp
