+++ src/PeakIntegrator.h
